build: add vercel config and update next.js config
Add vercel.json configuration file with build and runtime settings.
Update next.config.ts with experimental features, transpile packages, and webpack fallbacks for client-side bundling.
